Implemented TensorFlowObjectDetection Locator
