+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/modules/location/Locator.java
@@ -12,6 +12,7 @@ public interface Locator {
      * @see #isActive()
      * @see #getKind()
      */
+
     LocalizedMovement getLocation() throws LocatorException;
 
     /**
