+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleop/TeleOpMain.java
@@ -1,9 +1,14 @@
 package org.firstinspires.ftc.teamcode.teleop;
 
+import android.util.Size;
+
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.Gamepad;
 
+import org.firstinspires.ftc.robotcore.external.hardware.camera.BuiltinCameraDirection;
+import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.firstinspires.ftc.teamcode.OpBase;
+import org.firstinspires.ftc.vision.VisionPortal;
 
 import java.util.concurrent.atomic.AtomicBoolean;
 
@@ -12,9 +17,17 @@ public final class TeleOpMain extends OpBase {
 
     private Gamepad currentGamepad1, currentGamepad2, previousGamepad1, previousGamepad2;
 
+    private VisionPortal portal;
+
     @Override
     public void start() {
         super.start();
+
+        /*portal = new VisionPortal.Builder()
+                .setCamera(hardwareMap.get(WebcamName.class, "Webcam 1"))
+                .setCameraResolution(new Size(1280, 720))
+                .setCamera(BuiltinCameraDirection.BACK)
+                .build();*/
 
         previousGamepad1 = new Gamepad();
         previousGamepad2 = new Gamepad();
